VS2019 (#49)
* project adaptions for VS2019

* delete unused code

* bring to compile & run

* remove entries from context menu

* adapt build script

* fix msbuild path
