+++ IdeIntegration/Options/IntegrationOptions.cs
@@ -12,24 +12,12 @@
         public bool EnableStepMatchColoring { get; set; }
         public bool EnableTracing { get; set; }
         public string TracingCategories { get; set; }
-        public TestRunnerTool TestRunnerTool { get; set; }
         public bool DisableRegenerateFeatureFilePopupOnConfigChange { get; set; }
         public GenerationMode GenerationMode { get; set; }
         public string CodeBehindFileGeneratorPath { get; set; }
         public string CodeBehindFileGeneratorExchangePath { get; set; }
         public bool LegacyEnableSpecFlowSingleFileGeneratorCustomTool { get; set; }
         public bool OptOutDataCollection { get; set; }
-    }
-
-    public enum TestRunnerTool
-    {
-        Auto = 0,
-        ReSharper = 1,
-        VisualStudio2010MsTest = 2,
-        SpecRun = 3,
-        //TestDrivenDotNet = 4,
-        ReSharper5 = 5,
-        VisualStudio2012 = 6,
     }
 
     public enum GenerationMode
