+++ SpecFlow.VisualStudio.2019.sln
@@ -1,0 +1,81 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 16
+VisualStudioVersion = 15.0.27130.2010
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UnitTests", "UnitTests", "{E30EEB3E-0274-45EE-8FCE-4069823E670F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechTalk.SpecFlow.IdeIntegration", "IdeIntegration\TechTalk.SpecFlow.IdeIntegration.csproj", "{44D52C1B-0ABC-44B6-8EB8-0F754E264226}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechTalk.SpecFlow.IdeIntegration.UnitTests", "UnitTests\IdeIntegration.UnitTests\TechTalk.SpecFlow.IdeIntegration.UnitTests.csproj", "{69505BAC-12B9-4443-BB3B-A94E51EA5922}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechTalk.SpecFlow.VsIntegration.ItemTemplates", "ItemTemplates\TechTalk.SpecFlow.VsIntegration.ItemTemplates.csproj", "{662A06B2-A852-47C4-A258-D84FE81D4A80}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechTalk.SpecFlow.VsIntegration", "VsIntegration\TechTalk.SpecFlow.VsIntegration.csproj", "{96294E6F-8875-4D12-8577-4EC83E60BD6C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".Build", ".Build", "{DD083D4C-C52C-444C-8FEB-51FEFC77AEC1}"
+	ProjectSection(SolutionItems) = preProject
+		azure-pipelines.yml = azure-pipelines.yml
+		build.ps1 = build.ps1
+		build.yml = build.yml
+		changelog.txt = changelog.txt
+		CommonAssemblyInfo.cs = CommonAssemblyInfo.cs
+		GitVersion.yml = GitVersion.yml
+		LICENSE.txt = LICENSE.txt
+		publish.myget.ps1 = publish.myget.ps1
+		README.md = README.md
+		TechTalk.SpecFlow.VsIntegration.Common.Build.targets = TechTalk.SpecFlow.VsIntegration.Common.Build.targets
+		TechTalk.SpecFlow.VsIntegration.Common.targets = TechTalk.SpecFlow.VsIntegration.Common.targets
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechTalk.SpecFlow.RemoteAppDomain", "TechTalk.SpecFlow.RemoteAppDomain\TechTalk.SpecFlow.RemoteAppDomain.csproj", "{8B5402B9-7580-4AA3-9525-6C4563EF0623}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechTalk.SpecFlow.VisualStudio.CodeBehindGenerator", "TechTalk.SpecFlow.VisualStudio.CodeBehindGenerator\TechTalk.SpecFlow.VisualStudio.CodeBehindGenerator.csproj", "{5351E241-2B8B-4D04-8999-5E3A79E8DD64}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{93B345A8-B526-4B75-BF9A-930712FFC76E}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+		changelog.txt = changelog.txt
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{44D52C1B-0ABC-44B6-8EB8-0F754E264226}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44D52C1B-0ABC-44B6-8EB8-0F754E264226}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44D52C1B-0ABC-44B6-8EB8-0F754E264226}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44D52C1B-0ABC-44B6-8EB8-0F754E264226}.Release|Any CPU.Build.0 = Release|Any CPU
+		{69505BAC-12B9-4443-BB3B-A94E51EA5922}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{69505BAC-12B9-4443-BB3B-A94E51EA5922}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{69505BAC-12B9-4443-BB3B-A94E51EA5922}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{69505BAC-12B9-4443-BB3B-A94E51EA5922}.Release|Any CPU.Build.0 = Release|Any CPU
+		{662A06B2-A852-47C4-A258-D84FE81D4A80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{662A06B2-A852-47C4-A258-D84FE81D4A80}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{662A06B2-A852-47C4-A258-D84FE81D4A80}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{662A06B2-A852-47C4-A258-D84FE81D4A80}.Release|Any CPU.Build.0 = Release|Any CPU
+		{96294E6F-8875-4D12-8577-4EC83E60BD6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{96294E6F-8875-4D12-8577-4EC83E60BD6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{96294E6F-8875-4D12-8577-4EC83E60BD6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{96294E6F-8875-4D12-8577-4EC83E60BD6C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8B5402B9-7580-4AA3-9525-6C4563EF0623}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B5402B9-7580-4AA3-9525-6C4563EF0623}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B5402B9-7580-4AA3-9525-6C4563EF0623}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B5402B9-7580-4AA3-9525-6C4563EF0623}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5351E241-2B8B-4D04-8999-5E3A79E8DD64}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5351E241-2B8B-4D04-8999-5E3A79E8DD64}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5351E241-2B8B-4D04-8999-5E3A79E8DD64}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5351E241-2B8B-4D04-8999-5E3A79E8DD64}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{69505BAC-12B9-4443-BB3B-A94E51EA5922} = {E30EEB3E-0274-45EE-8FCE-4069823E670F}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FE82DEDC-FB35-4666-A342-E4BF059F1137}
+	EndGlobalSection
+EndGlobal
