+++ VsIntegration/16.0.source.extension.vsixmanifest
@@ -1,0 +1,29 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
+    <Metadata>
+        <Identity Id="TechTalk.SpecFlow.VisualStudioIntegration" Version="|%CurrentProject%;GetBuildVersion|" Language="en-US" Publisher="TechTalk" />
+        <DisplayName>SpecFlow for Visual Studio 2019</DisplayName>
+        <Description xml:space="preserve">SpecFlow integration for Visual Studio 2019. Sponsored by TechTalk (http://www.techtalk.at).</Description>
+        <MoreInfo>http://www.specflow.org</MoreInfo>
+        <License>LICENSE.txt</License>
+        <GettingStartedGuide>http://go.specflow.org/getting-started</GettingStartedGuide>
+        <ReleaseNotes>changelog.txt</ReleaseNotes>
+        <Icon>specflowsetup.ico</Icon>
+        <PreviewImage>specflow_preview.png</PreviewImage>
+        <Tags>BDD, specflow, gherkin, cucumber, living documentation, specification by example, atdd, Behavior Driven Development</Tags>
+    </Metadata>
+    <Installation>
+        <InstallationTarget Id="Microsoft.VisualStudio.Pro" Version="[15.0,17.0)" />
+    </Installation>
+    <Dependencies>
+        <Dependency Id="Microsoft.Framework.NDP" DisplayName="Microsoft .NET Framework" d:Source="Manual" Version="[4.6,)" />
+    </Dependencies>
+    <Assets>
+        <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="%CurrentProject%" Path="|%CurrentProject%|" />
+        <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="%CurrentProject%" Path="|%CurrentProject%;PkgdefProjectOutputGroup|" />
+        <Asset Type="Microsoft.VisualStudio.ItemTemplate" d:Source="Project" d:ProjectName="%CurrentProject%.ItemTemplates" d:TargetPath="|TechTalk.SpecFlow.VsIntegration.ItemTemplates;TemplateProjectOutputGroup|" Path="ItemTemplates" d:VsixSubPath="ItemTemplates" />
+    </Assets>
+    <Prerequisites>
+        <Prerequisite Id="Microsoft.VisualStudio.Component.Roslyn.LanguageServices" Version="[15.0,17.0)" DisplayName="C# and Visual Basic" />
+    </Prerequisites>
+</PackageManifest>
