+++ VsIntegration/Options/IntegrationOptionsProvider.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using EnvDTE;
 using Microsoft.VisualStudio.Shell;
 using TechTalk.SpecFlow.IdeIntegration.Options;
-using TechTalk.SpecFlow.VsIntegration.SingleFileGenerator;
 using TechTalk.SpecFlow.VsIntegration.Utils;
 
 namespace TechTalk.SpecFlow.VsIntegration.Options
@@ -26,7 +25,6 @@ namespace TechTalk.SpecFlow.VsIntegration.Options
         public const bool EnableStepMatchColoringDefaultValue = true;
         public const bool EnableTracingDefaultValue = false;
         public const string TracingCategoriesDefaultValue = "all";
-        public const TestRunnerTool TestRunnerToolDefaultValue = TestRunnerTool.Auto;
         public const bool DisableRegenerateFeatureFilePopupOnConfigChangeDefaultValue = false;
         public const GenerationMode GenerationModeDefaultValue = GenerationMode.OutOfProcess;
         public const string CodeBehindFileGeneratorPath = null;
@@ -34,7 +32,7 @@ namespace TechTalk.SpecFlow.VsIntegration.Options
         public const bool DefaultOptOutDataCollection = false;
 
 
-        private DTE _dte;
+        private DTE dte;
 
         public IntegrationOptionsProvider()
         {
@@ -42,7 +40,7 @@ namespace TechTalk.SpecFlow.VsIntegration.Options
 
         public IntegrationOptionsProvider(DTE dte)
         {
-            _dte = dte;
+            this.dte = dte;
         }
 
         private static T GetGeneralOption<T>(DTE dte, string optionName, T defaultValue = default(T))
@@ -50,7 +48,7 @@ namespace TechTalk.SpecFlow.VsIntegration.Options
             return VsxHelper.GetOption(dte, SPECFLOW_OPTIONS_CATEGORY, SPECFLOW_GENERAL_OPTIONS_PAGE, optionName, defaultValue);
         }
 
-        private IntegrationOptions GetOptions(DTE dte)
+        private static IntegrationOptions GetOptions(DTE dte)
         {
             var options = cachedOptions;
             if (options != null)
@@ -69,7 +67,6 @@ namespace TechTalk.SpecFlow.VsIntegration.Options
                 EnableStepMatchColoring = GetGeneralOption(dte, "EnableStepMatchColoring", EnableStepMatchColoringDefaultValue),
                 EnableTracing = GetGeneralOption(dte, "EnableTracing", EnableTracingDefaultValue),
                 TracingCategories = GetGeneralOption(dte, "TracingCategories", TracingCategoriesDefaultValue),
-                TestRunnerTool = GetGeneralOption(dte, "TestRunnerTool", TestRunnerToolDefaultValue),
                 DisableRegenerateFeatureFilePopupOnConfigChange = GetGeneralOption(dte, "DisableRegenerateFeatureFilePopupOnConfigChange", DisableRegenerateFeatureFilePopupOnConfigChangeDefaultValue),
                 GenerationMode = GetGeneralOption(dte, "GenerationMode", GenerationModeDefaultValue),
                 CodeBehindFileGeneratorPath = GetGeneralOption(dte, "PathToCodeBehindGeneratorExe", CodeBehindFileGeneratorPath),
@@ -80,22 +77,16 @@ namespace TechTalk.SpecFlow.VsIntegration.Options
             return options;
         }
 
-        private bool CheckIfSingleFileGeneratorIsEnabled(DTE dte)
-        {
-            var customToolSwitch = new CustomToolSwitch(dte);
-            return customToolSwitch.IsEnabled();
-        }
-
         [Import]
         internal SVsServiceProvider ServiceProvider
         {
-            set { _dte = VsxHelper.GetDte(value); }
+            set { dte = VsxHelper.GetDte(value); }
         }
 
 
         public IntegrationOptions GetOptions()
         {
-            return GetOptions(_dte);
+            return GetOptions(dte);
         }
     }
 }
