--- VsIntegration/TestRunner/ITestRunnerGatewayProvider.cs
@@ -1,9 +1,0 @@
-﻿using TechTalk.SpecFlow.IdeIntegration.Options;
-
-namespace TechTalk.SpecFlow.VsIntegration.TestRunner
-{
-    public interface ITestRunnerGatewayProvider
-    {
-        ITestRunnerGateway GetTestRunnerGateway(TestRunnerTool? runnerTool = null);
-    }
-}
