--- VsIntegration/TestRunner/SpecRunTestRunnerGatewayLoader.cs
@@ -1,11 +1,0 @@
-﻿using TechTalk.SpecFlow.IdeIntegration.Options;
-
-namespace TechTalk.SpecFlow.VsIntegration.TestRunner
-{
-    public class SpecRunTestRunnerGatewayLoader : SpecRunGatewayLoader
-    {
-        public SpecRunTestRunnerGatewayLoader() : base(TestRunnerTool.SpecRun)
-        {
-        }
-    }
-}
