--- VsIntegration/TestRunner/SpecRunWithVsTestExplorerGatewayLoader.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Linq;
-using EnvDTE;
-using TechTalk.SpecFlow.IdeIntegration.Options;
-using TechTalk.SpecFlow.VsIntegration.Utils;
-
-namespace TechTalk.SpecFlow.VsIntegration.TestRunner
-{
-    public class SpecRunWithVsTestExplorerGatewayLoader : SpecRunGatewayLoader
-    {
-        public SpecRunWithVsTestExplorerGatewayLoader() : base(TestRunnerTool.VisualStudio2012)
-        {
-        }
-    }
-}
