+++ azure-pipelines.yml
@@ -23,3 +23,13 @@ jobs:
     artifactFileName: '$(Build.ArtifactStagingDirectory)/SpecFlow.VisualStudio.VS2017.zip'
     pool:
       name: 'Hosted VS2017'
+
+- template: build.yml  
+  parameters:
+    name: VS2019
+    visualStudioVersion: '2019'
+    artifactName: 'SpecFlow.VisualStudio.VS2019.CI-BuildResult'
+    artifactFileName: '$(Build.ArtifactStagingDirectory)/SpecFlow.VisualStudio.VisualStudioIntegration.zip'
+    pool:
+      name: 'Hosted Windows 2019 with VS2019'
+
